+++ monarchy/src/main/java/com/zhuinden/monarchy/ManagedLiveResults.java
@@ -10,6 +10,7 @@ import io.realm.OrderedRealmCollectionChangeListener;
 import io.realm.Realm;
 import io.realm.RealmConfiguration;
 import io.realm.RealmModel;
+import io.realm.RealmQuery;
 import io.realm.RealmResults;
 
 /**
@@ -20,21 +21,23 @@ class ManagedLiveResults<T extends RealmModel>
         extends MutableLiveData<Monarchy.ManagedChangeSet<T>> {
     private final Monarchy monarchy;
     private final Monarchy.Query<T> query;
+    private final boolean asAsync;
 
     private final RealmConfiguration realmConfiguration;
 
     private OrderedRealmCollectionChangeListener<RealmResults<T>> realmChangeListener = new OrderedRealmCollectionChangeListener<RealmResults<T>>() {
         @Override
         public void onChange(@NonNull RealmResults<T> realmResults, @Nullable OrderedCollectionChangeSet changeSet) {
-            postValue(new Monarchy.ManagedChangeSet<>(realmResults, changeSet));
+            setValue(new Monarchy.ManagedChangeSet<>(realmResults, changeSet));
         }
     };
     private Realm realm;
     private RealmResults<T> realmResults;
 
-    public ManagedLiveResults(Monarchy monarchy, Monarchy.Query<T> query) {
+    public ManagedLiveResults(Monarchy monarchy, Monarchy.Query<T> query, boolean asAsync) {
         this.monarchy = monarchy;
         this.query = query;
+        this.asAsync = asAsync;
 
         this.realmConfiguration = this.monarchy.getRealmConfiguration();
     }
@@ -42,7 +45,12 @@ class ManagedLiveResults<T extends RealmModel>
     @Override
     protected void onActive() {
         realm = Realm.getInstance(realmConfiguration);
-        realmResults = query.createQuery(realm).findAllAsync(); // sort/distinct should be done with new queries
+        RealmQuery<T> realmQuery = query.createQuery(realm);
+        if(asAsync) {
+            realmResults = realmQuery.findAllAsync(); // sort/distinct should be done with new queries
+        } else {
+            realmResults = realmQuery.findAll(); // sort/distinct should be done with new queries
+        }
         realmResults.addChangeListener(realmChangeListener);
     }